refactor(GraphsScreen): extract hooks and modals - reduce from 1423 to 494 lines
Extract stateful logic into custom hooks:
- useExpenseData.js: expense data loading, aggregation, forecast calculations
- useIncomeData.js: income data loading and category aggregation
- useBalanceHistory.js: balance history data, trends, CRUD operations

Extract modal components:
- ChartModal.js: expense/income chart with category navigation
- BalanceHistoryModal.js: balance history table with inline editing

Benefits:
- 65% reduction (929 lines removed)
- Better separation of concerns
- Improved testability
- Code reuse potential
- Easier debugging and maintenance
